sources/client/requirements.txt
cx_Freeze==6.7
pyqt5-tools==5.15.4.3.2
sqlalchemy==1.3.24
sqlalchemy-utils==0.38.2
sqlacodegen==2.3.0
mysql-connector-python==8.0.28
pylint==2.12.2
qt_material==2.7